+++ Plugins/Sentry/Source/Sentry/Private/GenericPlatform/CrashReporter/GenericPlatformSentryCrashReporter.h
@@ -21,7 +21,9 @@ public:
 	void SetUser(TSharedPtr<FGenericPlatformSentryUser> user);
 	void RemoveUser();
 	void SetContext(const FString& key, const TMap<FString, FSentryVariant>& values);
+	void UpdateContext(const FString& key, const TMap<FString, FSentryVariant>& values);
 	void SetTag(const FString& key, const FString& value);
+	void SetTags(const TMap<FString, FString>& tags);
 	void RemoveTag(const FString& key);
 
 protected: